requirements.txt
django
django-rest-framework
pyyaml
django-cors-headers
requests


